Fix folder structure of the frontend
