--- unbrake-frontend/src/configuration/ConfigurationStyles.jsx
@@ -1,24 +1,0 @@
-const styles = theme => ({
-  container: {
-    display: "flex",
-    flexWrap: "wrap"
-  },
-  textField: {
-    marginLeft: theme.spacing.unit,
-    marginRight: theme.spacing.unit
-  },
-  dense: {
-    marginTop: 16
-  },
-  menu: {
-    width: 200
-  },
-  grid: {
-    padding: "px"
-  },
-  gridButton: {
-    paddingLeft: theme.spacing.unit + theme.spacing.unit
-  }
-});
-
-export default styles;
